Large header system refactoring
