+++ src/global/Main.c
@@ -6,7 +6,7 @@ int main( int argc, char *argv[] )
 {
         FireStream fst;
         FireStream_make(&fst);
-        FireStream_write_fmt(&fst, "blbl", 1, 300, 1, 300);
+        FireStream_int_push(&fst, 660);
         FireIO_print_binary_lines(&fst);
         FireIO_print_stream(&fst);
 }
